+++ pkg/core/hooks/conn/factory.go
@@ -5,11 +5,8 @@ package conn
 import (
 	"bytes"
 	"context"
-	"fmt"
 	"io"
 	"mime/multipart"
-	"net/http"
-	"net/url"
 	"strings"
 	"sync"
 	"time"
@@ -95,78 +92,6 @@ func (factory *Factory) GetOrCreate(connectionID ID) *Tracker {
 		return factory.connections[connectionID]
 	}
 	return tracker
-}
-
-func Capture(_ context.Context, logger *zap.Logger, t chan *models.TestCase, req *http.Request, resp *http.Response, reqTimeTest time.Time, resTimeTest time.Time, opts models.IncomingOptions) {
-	reqBody, err := io.ReadAll(req.Body)
-	if err != nil {
-		utils.LogError(logger, err, "failed to read the http request body")
-		return
-	}
-
-	defer func() {
-		err := resp.Body.Close()
-		if err != nil {
-			utils.LogError(logger, err, "failed to close the http response body")
-		}
-	}()
-
-	respBody, err := io.ReadAll(resp.Body)
-	if err != nil {
-		utils.LogError(logger, err, "failed to read the http response body")
-		return
-	}
-
-	if isFiltered(logger, req, opts) {
-		logger.Debug("The request is a filtered request")
-		return
-	}
-	var formData []models.FormData
-	if contentType := req.Header.Get("Content-Type"); strings.HasPrefix(contentType, "multipart/form-data") {
-		parts := strings.Split(contentType, ";")
-		if len(parts) > 1 {
-			req.Header.Set("Content-Type", strings.TrimSpace(parts[0]))
-		}
-		formData = extractFormData(logger, reqBody, contentType)
-		reqBody = []byte{}
-	} else if contentType := req.Header.Get("Content-Type"); contentType == "application/x-www-form-urlencoded" {
-		decodedBody, err := url.QueryUnescape(string(reqBody))
-		if err != nil {
-			utils.LogError(logger, err, "failed to decode the url-encoded request body")
-			return
-		}
-		reqBody = []byte(decodedBody)
-	}
-
-	t <- &models.TestCase{
-		Version: models.GetVersion(),
-		Name:    pkg.ToYamlHTTPHeader(req.Header)["Keploy-Test-Name"],
-		Kind:    models.HTTP,
-		Created: time.Now().Unix(),
-		HTTPReq: models.HTTPReq{
-			Method:     models.Method(req.Method),
-			ProtoMajor: req.ProtoMajor,
-			ProtoMinor: req.ProtoMinor,
-			// URL:        req.URL.String(),
-			// URL: fmt.Sprintf("%s://%s%s?%s", req.URL.Scheme, req.Host, req.URL.Path, req.URL.RawQuery),
-			URL: fmt.Sprintf("http://%s%s", req.Host, req.URL.RequestURI()),
-			//  URL: string(b),
-			Form:      formData,
-			Header:    pkg.ToYamlHTTPHeader(req.Header),
-			Body:      string(reqBody),
-			URLParams: pkg.URLParams(req),
-			Timestamp: reqTimeTest,
-		},
-		HTTPResp: models.HTTPResp{
-			StatusCode:    resp.StatusCode,
-			Header:        pkg.ToYamlHTTPHeader(resp.Header),
-			Body:          string(respBody),
-			Timestamp:     resTimeTest,
-			StatusMessage: http.StatusText(resp.StatusCode),
-		},
-		Noise: map[string][]string{},
-		// Mocks: mocks,
-	}
 }
 
 func extractFormData(logger *zap.Logger, body []byte, contentType string) []models.FormData {
